Get rear triangle dimensions

The rear triangle is made up of the projection in the horizontal plane of the chain stay, the seat stay, and the seat tube. The seat tube is the same as its projection in the horizontal plane, but the other two tubes stick out at an angle, ans they must accommodate the width of the rear wheel hub, so there's some trigonometry involved.

get_rear_triangle_dims(
  st_length = 500,
  cs_length = 450,
  st_angle = 71,
  cs_st_angle = 61,
  angle_btw_css = 14
)

Arguments

st_length
The length of the seat tube ST in millimeters.

cs_length
Chain stay length in millimeters.

st_angle
The angle between the seat tube and the horizontal.

cs_st_angle
Angle between chain stay and seat tube, in degrees.

angle_btw_css
Angle between chain stays, in degrees.

Value

A 3 x 3 tibble.

Details

A bike frame designer aims for a seat tube length (st_length) that fits a rider of a given height, and also for a seat tube angle (st_angle) and chain stay length (cs_length) that give the bike its desired riding characteristics, though cs_length is subject to some other constraints described in the documentation for get_cs_length().

The angle between the chain stay and the seat tube (cs_st_angle) is pre-determined in lugged frame construction by how the bottom bracket shell was cast.

This function returns the dimensions of the rear triangle using the convention described in the documentation for get_front_triangle_dims().

See also

Examples

get_rear_triangle_dims()
#> # A tibble: 3 × 3
#>   cs_triangle st_triangle ss_triangle
#>         <dbl>       <dbl>       <dbl>
#> 1       447.         500         483.
#> 2        77.6        473.        395.
#> 3       440.         163.        277.